update: CategoryView
