+++ PennyKeep/HomeView.swift
@@ -31,12 +31,7 @@ struct HomeView: View {
             .sorted { $0.date > $1.date }
     }
 
-    // Group recent transactions by category
-    private var recentTransactionsByCategory: [(category: String, transactions: [Transaction])] {
-        let grouped = Dictionary(grouping: recentTransactions) { $0.category }
-        return grouped.map { (category: $0.key, transactions: $0.value) }
-            .sorted { $0.category < $1.category }
-    }
+    
 
     // Computed array of totals for the last 12 months
     private var monthlyTotals: [(month: Date, income: Double, expense: Double)] {
@@ -85,42 +80,45 @@ struct HomeView: View {
                         .padding(.horizontal)
                     }
                     
-                    // Recent Transactions by Category
+                    // Recent Transactions List
                     VStack(alignment: .leading, spacing: 16) {
                         Text("Recent Transactions")
                             .font(.title2)
                             .bold()
                             .padding(.horizontal)
-                        
-                        ForEach(recentTransactionsByCategory, id: \.category) { group in
-                            VStack(alignment: .leading, spacing: 8) {
-                                Text(group.category)
-                                    .font(.headline)
-                                    .padding()
-                                
-                                ForEach(group.transactions) { transaction in
-                                    HStack {
-                                        VStack(alignment: .leading) {
-                                            Text(transaction.date, style: .date)
-                                                .font(.caption)
-                                            Text(transaction.title)
-                                                .font(.subheadline)
-                                                .bold()
-                                        }
-                                        Spacer()
-                                        let formattedAmount = transaction.amount.formatted(.currency(code: appSettings.selectedCurrency))
-                                        Text("\(transaction.type == .income ? "+" : "-")\(formattedAmount)")
+
+                        if recentTransactions.isEmpty {
+                            Text("No recent transactions")
+                                .foregroundColor(.secondary)
+                                .padding()
+                                .frame(maxWidth: .infinity, alignment: .center)
+                                .background(Color(UIColor.secondarySystemBackground))
+                                .cornerRadius(8)
+                                .padding(.horizontal)
+                        } else {
+                            ForEach(recentTransactions) { transaction in
+                                HStack {
+                                    VStack(alignment: .leading, spacing: 4) {
+                                        Text(transaction.title)
                                             .font(.headline)
-                                            .foregroundColor(transaction.type == .income ? .green : .red)
+                                        Text(transaction.category)
+                                            .font(.subheadline)
+                                            .foregroundColor(.secondary)
+                                        Text(transaction.date, style: .date)
+                                            .font(.caption)
+                                            .foregroundColor(.secondary)
                                     }
-                                    .padding(.vertical, 4)
-                                    .padding(.horizontal)
-                                    Divider()
+                                    Spacer()
+                                    let formattedAmount = transaction.amount.formatted(.currency(code: appSettings.selectedCurrency))
+                                    Text("\(transaction.type == .income ? "+" : "-")\(formattedAmount)")
+                                        .font(.headline)
+                                        .foregroundColor(transaction.type == .income ? .green : .red)
                                 }
+                                .padding()
+                                .background(Color(UIColor.secondarySystemBackground))
+                                .cornerRadius(8)
+                                .padding(.horizontal)
                             }
-                            .background(Color(UIColor.secondarySystemBackground))
-                            .cornerRadius(8)
-                            .padding(.horizontal)
                         }
                     }
                     .padding(.vertical)
